Generate some JSON data server-side
SEO engines start to indexing too early and the content we want to be indexed is missing... To avoid that, we make the first request API on the server-side and add it in the window object of JavaScript. So we can put the values immediatly in the Redux store and get the content displayed before AJAX requests.