Download npm packages in temporary folder
Before, we copy the project folder and then download the NPM packages. This way, the packages were always download which is painful in a slow network environment.
To avoid that, we download NPM packages in a temporary folder and them we link to that folder. So, we take profit of the Docker layer mecanism to avoid downloads at each deploy.