Include polyfill for GoogleBot + Reduce companies JSON size
When using 'fetch as Google' in the Google Search Console, we got a white page. Based on this article https://medium.com/@andrewzey/google-seo-with-create-react-app-fixing-the-hidden-gotcha-c164063106d, we include the babel-polyfill in the index page - and it works.
At the same time, we reduce the JSON of the companies included in the HTML template.