AB test the fact of showing company offers when present For group A, no visible change.
For group B, we show a label on the result list to differentiate
which companies have offers and which do not. And on the detail page
of each company we show direct links to its offers if present.

Existing GA pageviews (company detail and show contact info buttons)
now trigger additional pageviews to differenciate companies with offers
and without offers.

Moreover, the AB testing is materialized in GA as an additional GA event.

This way for example for company detail events, we now have 4 metrics:
one for each combination of (group A, group B) [materialized as GA events]
and (company has offers or not) [materialized as GA pageviews].